pyproject.toml
[project]
name = "lingnexus"
version = "0.1.0"
description = "基于 AgentScope 的多智能体系统，支持 Claude Skills 兼容"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agentscope>=1.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "python-docx>=1.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

# 如果不需要构建为包，可以移除 build-system 配置
# 但 uv sync 仍然需要它来管理依赖
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lingnexus"]  # 必须与项目名称匹配，否则构建失败

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312', 'py313']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # 行长度由 black 处理

